Fix `AudioFileWaveforms` not showing waves sometimes.
The problem is that this widget sometimes is being initialized before the player controller could get
the the buffer data, so this widget would be initialized with an empty wave data.
So we add a listner to the player to watch it's state, when it's properly finished reading the audio
file, we will recalculate the wave.